Migliorie generali

+++ src/main/java/com/example/warehouse/controller/AutomationRestController.java
@@ -9,6 +9,7 @@ import com.example.warehouse.service.UserRestService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -30,16 +31,18 @@ public class AutomationRestController {
         return new ReturnWithMessage(true, "", automationRestService.getAutomations());
     }*/
     @GetMapping("/discount/{id}")
-    public void DiscountManagement(@PathVariable("id") long id){
+    public void DiscountManagement(@PathVariable("id") long id) {
         automationRestService.DiscountManagement(id);
     }
 
     @PostMapping("/auto")
-    public List<Automation> list(@RequestBody Map<String, String> request){
+    public List<Automation> list(@RequestBody Map<String, String> request) {
         //automationRestService.getAutomations();
         String token = request.get("userToken").toString();
         System.out.println("token: " + token);
-        return automationRestService.auto();
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.auto();
+        return new ArrayList<>();
     }
 
     @PostMapping("/discountRule/add")
@@ -48,10 +51,12 @@ public class AutomationRestController {
         int period = Integer.parseInt(request.get("period").toString());
         int typePeriod = Integer.parseInt(request.get("typePeriod").toString());
         String token = request.get("userToken").toString();
-
-        ReturnWithMessage r = automationRestService.addDiscountRule(new DiscountRules(percentage, period, typePeriod));
-        System.out.println(r);
-        return r;
+        if (automationRestService.checkToken(token, 15)) {
+            ReturnWithMessage r = automationRestService.addDiscountRule(new DiscountRules(percentage, period, typePeriod));
+            System.out.println(r);
+            return r;
+        }
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/stockRule/add")
@@ -59,14 +64,19 @@ public class AutomationRestController {
         String token = request.get("userToken").toString();
         boolean minMax = Boolean.getBoolean(request.get("minMax").toString());
         long number = Integer.parseInt(request.get("number").toString());
-        return automationRestService.addStockRule(new StocksRules(minMax, number));
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.addStockRule(new StocksRules(minMax, number));
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
+
     @PostMapping("/stockRule/modify")
     public ReturnWithMessage modifyStockRule(@RequestBody Map<String, Object> request) {
         String token = request.get("userToken").toString();
         long number = Long.parseLong(request.get("number").toString());
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.modifyStockRule(number, id);
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.modifyStockRule(number, id);
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/discountRule/modify")
@@ -76,39 +86,50 @@ public class AutomationRestController {
         int period = Integer.parseInt(request.get("period").toString());
         int typePeriod = Integer.parseInt(request.get("typePeriod").toString());
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.modifyDiscountRule(new DiscountRules(id, percentage, period, typePeriod));
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.modifyDiscountRule(new DiscountRules(id, percentage, period, typePeriod));
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/stockRule/delete")
     public ReturnWithMessage removeStockRule(@RequestBody Map<String, Object> request) {
         String token = request.get("userToken").toString();
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.removeStockRule(id);
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.removeStockRule(id);
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/discountRule/delete")
     public ReturnWithMessage removeDiscountRule(@RequestBody Map<String, Object> request) {
         String token = request.get("userToken").toString();
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.removeDiscountRule(id);
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.removeDiscountRule(id);
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/automation/active")
     public ReturnWithMessage activeAutomation(@RequestBody Map<String, Object> request) {
         String token = request.get("userToken").toString();
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.activeAutomation(id);
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.activeAutomation(id);
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
     @PostMapping("/automation/disable")
     public ReturnWithMessage disableAutomation(@RequestBody Map<String, Object> request) {
         String token = request.get("userToken").toString();
         long id = Long.parseLong(request.get("id").toString());
-        return automationRestService.disableAutomation(id);
+        if (automationRestService.checkToken(token, 15))
+            return automationRestService.disableAutomation(id);
+        return new ReturnWithMessage(false, "Non hai i permessi");
     }
 
+
     @GetMapping("/automation/management/{id}")
-    public void OrderManagement(@PathVariable("id") long idProduct){
+    public void OrderManagement(@PathVariable("id") long idProduct) {
         automationRestService.OrderManagement(idProduct);
     }
